Allow adding libraries via postMessage (#74)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This change adds a postMessage API for adding libraries, this is primarily for Eugene Gavriloffs use in Figma2FAST.

### 🎫 Issues

<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->
Closes #42

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
The logic has been refactored to a class that adds the window message listener and checks for multiple kinds of postMessage. It uses the legacy data-dictionary update and this has been marked as such. As tests have still not been set up for this repository, this has been tested manually only using the devtools to send a message to the creator, you can try this yourself by using the following command in the devtools console:

`window.postMessage("{\"type\": \"library::add\", \"libraryId\": \"fast-components\"}");`

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [x] I have updated the project documentation to reflect my changes.