Fixed the issue with upload/download by ensuring all JSON schemas are added during initialization (#89)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This fixes the upload issue where if the project file included any libraries, the schemas for these components were not added and caused runtime issues.

### 🎫 Issues

<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->
Closes #29

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
Add a component (let's say fluent) to the document, download the project file. Refresh the page and attempt to upload the project file. This should result in the correct components rendering with no runtime errors.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.