Add css box model UI (#76)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
Replacing the margin, border and width fields in the editor pane with the new CSS Box Model component. This new UI will allow for the modification of margin, border widths, padding, height and width styles with a more visually intuitive UI. 
This also increases the width of the edit pane to 470px in order to fit the box model UI.

**Note: This will not pass the build gate until after Creator's fast-tooling dependency version is updated to include the new CSS Box Model component.**

### 🎫 Issues
Closes: #25 
<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->

Co-authored-by: Jane Chu <7559015+janechu@users.noreply.github.com>