Add workflow and scripting to automatically check and update the stage branch when FAST Tooling dependencies are updated (#100)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This automated job rubs 1 hour after the FAST Tooling publish. It will check to see if the versions have been updated and if so will update them and run the script to generate new changelog versions for the "What's New" feature.

This should cut down on the possibility that these changes are not added correctly, and that they can now be done in an automated fashion. I have also enabled this for dispatch should it be necessary.

### 🎫 Issues

<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->
Closes #91

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [x] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->
- See if this works, the individual pieces have been tested however the full workflow will need to be checked after packages have been released from FAST Tooling