Move the assets to the app folder root and remove the unused assets, keeping the generated assets in case they are needed (#107)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This change moves the site utilities assets to the app root assets folder. It removes unused files that are related to FAST and not FAST Creator.

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
Check the site renders correctly, none of these assets should be in use as they were sent to a CDN to be referenced except for the favicon.ico file. That should still work correctly.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.