Remove the message system setup logic from the preview component (#131)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
As part of the effort to implement shortcuts, the message system logic for the preview component has been moved external to the component. This should facilitate both the inclusion of shortcuts, and the packaging up of the preview web component as part of FAST tooling at a later date.

### 🎫 Issues

<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->
Begins to resolve #120

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->
- Add shortcuts to the preview